Update Rails to v7.0; remove oracle
There's apparently an incompatibility between Rails 6.x and Ruby 3.2:

"
NameError: uninitialized constant ActiveSupport::LoggerThreadSafeLevel::Logger

"This is a known issue when using Ruby 3.2+ with Rails 6.1.
Rails 6.1 is not fully compatible with Ruby 3.2. The LoggerThreadSafeLevel::Logger constant was removed from Ruby's standard library, but Rails 6.1 still references it."

Upgrading to Rails 7 fixed the issue I was having with RSpec not running.

I don't think we've needed the Oracle dependency for a while. Faced with trying to recompile a new
version for native extensions after updating Rails, this seemed like an opportune time to finally
get rid of it.